+++ src/test/java/com/umust/dobonglife/domain/mypage/controller/MyPageControllerTest.java
@@ -37,6 +37,8 @@ import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.restdocs.request.RequestDocumentation.*;
+import static com.epages.restdocs.apispec.ResourceDocumentation.parameterWithName;
+import static com.epages.restdocs.apispec.SimpleType.INTEGER;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -113,7 +115,7 @@ class MyPageControllerTest {
                                             .description("마이페이지 정보와 포인트 내역을 조회합니다.")
                                             .queryParameters(
                                                     parameterWithName("size").optional()
-                                                            .description("포인트 내역 조회 개수 (기본값: 3)")
+                                                            .description("포인트 내역 조회 개수 (기본값: 3)").type(INTEGER)
                                             )
                                             .responseFields(
                                                     fieldWithPath("success").type(JsonFieldType.BOOLEAN).description("요청 성공 여부"),
@@ -203,9 +205,9 @@ class MyPageControllerTest {
                                             .description("찜한 코스 목록을 조회합니다.")
                                             .queryParameters(
                                                     parameterWithName("lastId").optional()
-                                                            .description("커서 - 마지막 코스 ID (첫 요청 시 생략)"),
+                                                            .description("커서 - 마지막 코스 ID (첫 요청 시 생략)").type(INTEGER),
                                                     parameterWithName("size").optional()
-                                                            .description("조회 개수 (기본값: 3)")
+                                                            .description("조회 개수 (기본값: 3)").type(INTEGER)
                                             )
                                             .responseFields(
                                                     fieldWithPath("success").type(JsonFieldType.BOOLEAN).description("요청 성공 여부"),
@@ -297,9 +299,9 @@ class MyPageControllerTest {
                                             .description("찜한 장소 목록을 조회합니다.")
                                             .queryParameters(
                                                     parameterWithName("lastId").optional()
-                                                            .description("커서 - 마지막 장소 ID (첫 요청 시 생략)"),
+                                                            .description("커서 - 마지막 장소 ID (첫 요청 시 생략)").type(INTEGER),
                                                     parameterWithName("size").optional()
-                                                            .description("조회 개수 (기본값: 3)")
+                                                            .description("조회 개수 (기본값: 3)").type(INTEGER)
                                             )
                                             .responseFields(
                                                     fieldWithPath("success").type(JsonFieldType.BOOLEAN).description("요청 성공 여부"),
